moved armor stand related events to separate class
